# src/generated/service.openapi.json
{"openapi":"3.0.0","info":{"title":"BFF_Client_API","description":"Service to provide the communication between frontend functionality and the database.","version":"1.0.0"},"servers":[{"url":"https://api.eu-central-1.dev.services.vwfs.io/BFF_Client_API","description":"Dev server"},{"url":"https://api.eu-central-1.int.services.vwfs.io/BFF_Client_API","description":"Int server"},{"url":"https://api.eu-central-1.cons.services.vwfs.io/BFF_Client_API","description":"cons server"},{"url":"https://api.eu-central-1.services.vwfs.io/BFF_Client_API","description":"Production server"}],"security":[{"basicAuth":[]}],"paths":{"/triggerJob":{"get":{"tags":["API"],"operationId":"triggerJob","parameters":[{"in":"query","name":"jobName","required":true,"schema":{"type":"string"}},{"in":"query","name":"jobType","required":true,"schema":{"type":"string"}},{"in":"header","name":"tenantid","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Sync Job Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/triggerJob_200"}}}},"201":{"description":"Async Job Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/triggerJob_201"}}}},"404":{"description":"Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/triggerJob_404"}}}}}}},"/getAllObjects/{tenantId}":{"get":{"tags":["API"],"operationId":"getAllObjects","parameters":[{"in":"path","name":"tenantId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of objects","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/getObjectFields/{tenantId}/{objectName}":{"get":{"tags":["API"],"operationId":"getObjectFields","parameters":[{"in":"path","name":"tenantId","required":true,"schema":{"type":"string"}},{"in":"path","name":"objectName","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"object fields","content":{"application/json":{"schema":{"$ref":"#/components/schemas/object_fields_200"}}}}}}},"/createEndpoint":{"post":{"tags":["API"],"operationId":"createEndpoint","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"method":{"type":"integer","enum":[1,2]},"sourceId":{"type":"string"},"json":{"type":"string"}}}}}},"responses":{"200":{"description":"create endpoint","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}}}}}}}}},"/createSFConnection":{"post":{"tags":["API"],"operationId":"createSFConnection","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"tenantId":{"type":"string"},"sourceName":{"type":"string"},"clientSecret":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"baseUrl":{"type":"string"},"clientId":{"type":"string"}}}}}},"responses":{"200":{"description":"create SF Connection","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}}}}}}}}},"/getSFObjectChanges/{tenantId}":{"get":{"tags":["API"],"operationId":"getSFObjectChanges","parameters":[{"in":"path","name":"tenantId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"get SF Object Changes","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getSFObjectChanges_200"}}}}}}},"/data/{endpointName}":{"get":{"tags":["API"],"operationId":"data","parameters":[{"in":"path","name":"endpointName","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"get dynamic endpoint response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}}}},"/addTemplate":{"post":{"tags":["API"],"operationId":"addTemplate","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"excelFile":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"add Template","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}}}}}}}}}},"components":{"securitySchemes":{"basicAuth":{"type":"http","scheme":"basic"}},"schemas":{"triggerJob_200":{"type":"object","properties":{"totalInputRecords":{"type":"integer"},"totalProcessed":{"type":"integer"},"success":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"success":{"type":"boolean"},"errors":{"type":"array","items":{"type":"string"}},"created":{"type":"boolean"}}}},"failure":{"type":"array","items":{"type":"object","properties":{"status":{"type":"number"},"message":{"type":"string"}}}}}},"triggerJob_201":{"type":"object","properties":{"totalInputRecords":{"type":"integer"},"totalProcessed":{"type":"integer"},"jobId":{"type":"string"}}},"triggerJob_404":{"type":"object","properties":{"status":{"type":"integer"},"message":{"type":"string"}}},"object_fields_200":{"type":"object","properties":{"fields":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"type":{"type":"string"}}}},"childRelationships":{"type":"array","items":{"type":"object","properties":{"childSObject":{"type":"string"},"fields":{"type":"string"}}}}}},"getSFObjectChanges_200":{"type":"object","properties":{"data":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"tenantId":{"type":"string"},"deltas":{"type":"string"},"changesCount":{"type":"number"},"createdAt":{"type":"string","format":"date-time"}}}}}}}}}